VS 2015/17/19 update
1. Code refreshed for new C++ standarts [support VS 2015/2017/2019]
2. Fixed audio engine bug (missing microphone crash)
3. Fixed minor issue with loading UI screen (missing update events)
4. Remove all project warnings.
5. Added _MSC_PLATFORM_TOOLSET macro in project settings.
6. Added "/bigobj" to project C/C++ command line

+++ SVFEngine/SVFEngine/RenderDX9.h
@@ -260,7 +260,7 @@ namespace SAVFGAME
 							break;
 						else
 						{
-							if (!(i % 15)) printf("\nRENDER : Can't reset D3D device! (%i sec.)", i);
+							if (!(i % 15)) printf("\nRENDER : Can't reset D3D device! (%i sec.)", (int)i);
 						}
 					}
 				}
